+++ R/plot.p_significance.R
@@ -90,7 +90,7 @@ data_plot.p_significance <- function(x, data = NULL, grid = TRUE, show_intercept
     ylab <- "Parameters"
   }
 
-
+  dataplot <- .fix_facet_names(dataplot)
 
   attr(dataplot, "info") <- list("xlab" = "Possible parameter values",
                                  "ylab" = ylab,
@@ -174,11 +174,13 @@ plot.see_p_significance <- function(x, data = NULL, show_intercept = FALSE, prio
     geom_vline(aes(xintercept = 0), linetype = "dotted") +
     guides(fill = FALSE, color = FALSE, group = FALSE)
 
-  if ("Effects" %in% names(x) && isTRUE(grid)) {
-    if ("Component" %in% names(x))
+  if (isTRUE(grid)) {
+    if ("Component" %in% names(x) && "Effects" %in% names(x))
       p <- p + facet_wrap(~ Effects + Component, scales = "free")
-    else
+    else if ("Effects" %in% names(x))
       p <- p + facet_wrap(~ Effects, scales = "free")
+    else if ("Component" %in% names(x))
+      p <- p + facet_wrap(~ Component, scales = "free")
   }
 
   p