+++ R/utils.R
@@ -85,3 +85,17 @@ magrittr::`%>%`
 
   params
 }
+
+
+
+.fix_facet_names <- function(x) {
+  if ("Component" %in% names(x)) {
+    x$Component[x$Component == "conditional"] <- "(Conditional)"
+    x$Component[x$Component == "zero_inflated"] <- "(Zero-Inflated)"
+  }
+  if ("Effects" %in% names(x)) {
+    x$Effects[x$Effects == "fixed"] <- "Fixed Effects"
+    x$Effects[x$Effects == "random"] <- "Random Effects"
+  }
+  x
+}