+++ NEWS.md
@@ -18,7 +18,7 @@
 * `plot()` for `parameters::model_parameters()` now also create forest plots for meta-analysis.
 * `plot()` for `bayestestR::bayesfactor_parameters()` only plots facets when necessary.
 * `plot()` for `performance::check_outliers()` now also plot multiple methods in one plot.
-* Following `plot()` methods get a `grid`-argument, so model components like random effects or a zero-inflation component are plotted in a grid-layout: `bayestestR::p_direction()`, `bayestestR::hdi()`, `bayestestR::rope()`
+* Following `plot()` methods get a `grid`-argument, so model components like random effects or a zero-inflation component are plotted in a grid-layout: `bayestestR::p_direction()`, `bayestestR::hdi()`, `bayestestR::rope()`, `bayestestR::equivalence_test()`
 * Following `plot()` methods get `priors` and `priors_alpha` arguments, to add a layer with prior samples to the plot: `bayestestR::p_direction()`, `bayestestR::estimate_density()`, `bayestestR::point_estimate()`
 * Improved layout (vertical alignment) for some plotting-functions for the **bayestestR**-package, when there was only one parameter. Plots now are closer to the x-axis, removing unnecessary empty space.